formatting, I think I'm done before I mess something up. It's getting out of hand

+++ pydfnworks/pydfnworks/dfnGen/meshing/run_meshing.py
@@ -14,6 +14,7 @@ from shutil import copy, rmtree
 from numpy import genfromtxt
 from pydfnworks.dfnGen.meshing import mesh_dfn_helper as mh
 
+
 def mesh_fracture(fracture_id, visual_mode, num_poly):
     """Child function for parallelized meshing of fractures
 
@@ -38,9 +39,12 @@ def mesh_fracture(fracture_id, visual_mode, num_poly):
 
     """
 
-    t = time.time()
+    tic = time.time()
     p = mp.current_process()
-    print(f"--> Fracture {fracture_id} \tstarting on {p.name}")
+    digits = len(str(num_poly))
+    print(
+        f"--> Fracture {fracture_id:0{digits}d} out of {num_poly} is starting on {p.name}"
+    )
 
     _, cpu_id = p.name.split("-")
     cpu_id = int(cpu_id)
@@ -55,9 +59,13 @@ def mesh_fracture(fracture_id, visual_mode, num_poly):
         os.symlink(f"intersections/intersections_{fracture_id}.inp",\
             f"intersections_CPU{cpu_id}.inp")
 
-        os.symlink(f"points/points_{fracture_id}.xyz",f"points_CPU{cpu_id}.xyz")
+        os.symlink(f"points/points_{fracture_id}.xyz",
+                   f"points_CPU{cpu_id}.xyz")
 
-    mh.run_lagrit_script(f"mesh_poly_CPU{cpu_id}.lgi",output_file=f"lagrit_logs/log_lagrit_{fracture_id}",quite=True)
+    mh.run_lagrit_script(
+        f"mesh_poly_CPU{cpu_id}.lgi",
+        output_file=f"lagrit_logs/log_lagrit_{fracture_id:0{digits}d}.out",
+        quite=True)
 
     if not visual_mode:
         ## Once meshing is complete, check if the lines of intersection are in the final mesh
@@ -66,10 +74,11 @@ def mesh_fracture(fracture_id, visual_mode, num_poly):
             id_tri_node_CPU{cpu_id}.list \
             mesh_{fracture_id}.inp \
             {fracture_id}"
+
         failure = subprocess.call(cmd_check, shell=True)
 
         # If the lines of intersection are not in the final mesh, put depending files
-        # into a directory for debugging. 
+        # into a directory for debugging.
         if failure:
             print(f"--> WARNING: MESH CHECKING FAILED on {fracture_id}!!!!")
 
@@ -86,14 +95,17 @@ def mesh_fracture(fracture_id, visual_mode, num_poly):
             copy(f"{fracture_id}_mesh_errors.txt", folder)
             copy(f"poly_CPU{cpu_id}.inp", folder)
             copy(f"id_tri_node_CPU{cpu_id}.list", folder)
-            copy(f"intersections_CPU{cpu_id}.inp", folder )
-            copy(f"lagrit_logs/log_lagrit_{fracture_id}", folder)
+            copy(f"intersections_CPU{cpu_id}.inp", folder)
+            copy(f"lagrit_logs/log_lagrit_{fracture_id:0{digits}d}.out",
+                 folder)
             copy(f"parameters_CPU{cpu_id}.mlgi", folder)
             copy(f"mesh_poly_CPU{cpu_id}.lgi", folder)
 
             ## remove links
-            files = [f'poly_CPU{cpu_id}.inp', f'intersections_CPU{cpu_id}.inp', 
-                f'points_CPU{cpu_id}.xyz',f'parameters_CPU{cpu_id}.mlgi']
+            files = [
+                f'poly_CPU{cpu_id}.inp', f'intersections_CPU{cpu_id}.inp',
+                f'points_CPU{cpu_id}.xyz', f'parameters_CPU{cpu_id}.mlgi'
+            ]
             for f in files:
                 try:
                     os.unlink(f)
@@ -115,18 +127,23 @@ def mesh_fracture(fracture_id, visual_mode, num_poly):
 
     # Remove old links and files
     if visual_mode:
-        files = [f'poly_CPU{cpu_id}.inp',f'parameters_CPU{cpu_id}.mlgi']
+        files = [f'poly_CPU{cpu_id}.inp', f'parameters_CPU{cpu_id}.mlgi']
     else:
-        files = [f'poly_CPU{cpu_id}.inp', f'intersections_CPU{cpu_id}.inp', 
-            f'points_CPU{cpu_id}.xyz',f'parameters_CPU{cpu_id}.mlgi']
+        files = [
+            f'poly_CPU{cpu_id}.inp', f'intersections_CPU{cpu_id}.inp',
+            f'points_CPU{cpu_id}.xyz', f'parameters_CPU{cpu_id}.mlgi'
+        ]
     for f in files:
         try:
             os.unlink(f)
         except:
             print(f'Warning: Could unlink {f}')
 
-    elapsed = time.time() - t
-    print(f"--> Meshing fracture {fracture_id} out of {num_poly} complete. Time required: {elapsed:.2f} seconds\n")
+    elapsed = time.time() - tic
+
+    print(
+        f"--> Fracture {fracture_id:0{digits}d} out of {num_poly} is complete on {p.name}. Time required: {elapsed:.2f} seconds\n"
+    )
 
 
 def single_worker(work_queue, visual_mode, num_poly):
@@ -189,12 +206,14 @@ def mesh_fractures_header(fracture_list, ncpu, visual_mode):
 
     """
     t_all = time.time()
+    print()
     print('=' * 80)
-    print(f"\n--> Triangulating {len(fracture_list)} fractures using {ncpu} CPUS")
-    try:
+    print(
+        f"\n--> Triangulating {len(fracture_list)} fractures using {ncpu} processors\n\n"
+    )
+
+    if os.path.isdir('lagrit_logs'):
         rmtree('lagrit_logs')
-    except OSError:
-        pass
     os.mkdir('lagrit_logs')
 
     # create failure log file
@@ -219,7 +238,7 @@ def mesh_fractures_header(fracture_list, ncpu, visual_mode):
     for p in processes:
         p.join()
 
-    elapsed = (time.time() - t_all) /60.
+    elapsed = (time.time() - t_all)
 
     if os.stat("failure.txt").st_size > 0:
         failure_list = genfromtxt("failure.txt")
@@ -233,9 +252,15 @@ def mesh_fractures_header(fracture_list, ncpu, visual_mode):
         failure_flag = False
         os.remove("failure.txt")
         print('--> Triangulating Polygons: Complete')
-        print(f"--> Total Time to Mesh Network: {elapsed:.2f} minutes")
+
+        if elapsed > 60:
+            print(f"--> Total Time to Mesh Network: {elapsed:.2f} minutes")
+        else:
+            print(f"--> Total Time to Mesh Network: {elapsed:.2f} seconds")
+
         print('=' * 80)
     return failure_flag
+
 
 def merge_worker(job):
     """Parallel worker for merge meshes into final mesh 
@@ -256,13 +281,17 @@ def merge_worker(job):
     print(f"--> Starting merge: {job}")
     tic = time.time()
 
-    if mh.run_lagrit_script(f"merge_poly_part_{job}.lgi",f"lagrit_logs/log_merge_poly_part{job}",quite=True):
+    if mh.run_lagrit_script(f"merge_poly_part_{job}.lgi",
+                            f"lagrit_logs/log_merge_poly_part{job}.out",
+                            quite=True):
         print(f"Error {job} failed")
         return True
 
     toc = time.time()
     elapsed = toc - tic
-    print(f"--> Merge Number {job} Complete. Time elapsed: {elapsed:.2f} seconds")
+    print(
+        f"--> Merge Number {job} Complete. Time elapsed: {elapsed:.2f} seconds"
+    )
     return False
 
 
@@ -306,31 +335,16 @@ def merge_the_meshes(num_poly, ncpu, n_jobs, visual_mode):
             sys.stderr.write(error)
             sys.exit(1)
 
-    # for j in range(1, n_jobs + 1):
-    #     pid = os.fork()
-    #     if pid == 0: # clone a child job
-    #         cmd = os.environ['LAGRIT_EXE']+ ' < merge_poly_part_%d.lgi ' \
-    #             + '> log_merge_poly_part%d'
-    #         subprocess.call(cmd%(j,j), shell = True)
-    #         os._exit(0)
-    #     else:
-    #         print('Merging part ', j, ' of ', n_jobs)
-
-    # # wait for all child processes to complete
-    # j = 0
-    # while j < n_jobs:
-    #     (pid, status) = os.waitpid(0,os.WNOHANG)
-    #     if pid > 0:
-    #         print('Process ' + str(j+1) + ' finished')
-    #         j += 1
     print('=' * 80)
     print("--> Starting Final Merge")
     tic = time.time()
-    mh.run_lagrit_script('merge_rmpts.lgi','lagrit_logs/log_merge_all.txt',quite=True)
+    mh.run_lagrit_script('merge_rmpts.lgi',
+                         'lagrit_logs/log_merge_all.out',
+                         quite=True)
     toc = time.time()
     elapsed = toc - tic
     print(f"--> Final merge took {elapsed:.2f} seconds")
-    # Check log_merge_all.txt for LaGriT complete successfully
+
     if not visual_mode:
         if (os.stat("full_mesh.lg").st_size > 0):
             print("--> Final merge successful")
@@ -339,6 +353,7 @@ def merge_the_meshes(num_poly, ncpu, n_jobs, visual_mode):
             error = "ERROR: Final merge Failed\n"
             sys.stderr.write(error)
             sys.exit(1)
+
     else:
         if os.stat("reduced_mesh.inp").st_size > 0:
             print("--> Final merge successful")
